fix: build error

+++ src/MeoAssistant/InfrastProductionTask.cpp
@@ -471,7 +471,7 @@ bool asst::InfrastProductionTask::optimal_calc()
         if (cur_combs.size() < cur_max_num_of_opers) {
             // 允许外部的话，就把单个干员凑进来
             if (group.allow_external) {
-                int substitutes = cur_max_num_of_opers - cur_combs.size();
+                size_t substitutes = cur_max_num_of_opers - cur_combs.size();
                 for (size_t index = 0; index < substitutes; ++index) {
                     const auto& comb = cur_available_opers.at(index);
                     cur_combs.emplace_back(comb);
